Factor out ControllerMapping interface and subclasses
