+++ dji_robomaster_ep_driver/src/joystick_teleop_node.ros1.cpp
@@ -1,3 +1,6 @@
+#include <dji_robomaster_ep_driver/joystick_controller_mappings.hpp>
+
+#include <memory>
 #include <stdexcept>
 #include <string>
 
@@ -13,19 +16,39 @@ class ControllerTeleop
 {
 public:
   ControllerTeleop(
-      const ros::NodeHandle& nh, const std::string& joy_topic,
-      const std::string& command_topic, const std::string& command_frame,
-      const double max_linear_velocity, const double max_angular_velocity)
-      : nh_(nh), max_linear_velocity_(max_linear_velocity),
-        max_angular_velocity_(max_angular_velocity)
+      const ros::NodeHandle& nh, const std::string& joy_type,
+      const std::string& joy_topic, const std::string& command_topic,
+      const std::string& command_frame, const double max_linear_velocity,
+      const double max_angular_velocity)
+      : nh_(nh)
   {
-    if (max_linear_velocity_ <= 0.0)
+    if (max_linear_velocity <= 0.0)
     {
       throw std::invalid_argument("max_linear_velocity_ <= 0.0");
     }
-    if (max_angular_velocity_ <= 0.0)
+    if (max_angular_velocity <= 0.0)
     {
       throw std::invalid_argument("max_angular_velocity_ <= 0.0");
+    }
+
+    if (joy_type == "xbox_one")
+    {
+      ROS_INFO("Configured teleop controller for Xbox One controller");
+      controller_mapping_ = std::unique_ptr<ControllerMapping>(
+          new XboxOneControllerMapping(max_linear_velocity,
+                                       max_angular_velocity));
+    }
+    else if (joy_type == "3d_pro")
+    {
+      ROS_INFO("Configured teleop controller for Logitech 3D PRO controller");
+      controller_mapping_ = std::unique_ptr<ControllerMapping>(
+          new Logitech3DProControllerMapping(max_linear_velocity,
+                                             max_angular_velocity));
+    }
+    else
+    {
+      throw std::invalid_argument(
+          "Joystick type [" + joy_type + "] not supported");
     }
 
     velocity_command_.header.frame_id = command_frame;
@@ -48,18 +71,11 @@ public:
     }
   }
 
-  double max_linear_velocity() const { return max_linear_velocity_; }
-
-  double max_angular_velocity() const { return max_angular_velocity_; }
-
-protected:
-  virtual geometry_msgs::Twist ComputeVelocityCommand(
-      const sensor_msgs::Joy& joy_msg) const = 0;
-
 private:
   void JoyCallback(const sensor_msgs::Joy& joy_msg)
   {
-    velocity_command_.twist = ComputeVelocityCommand(joy_msg);
+    velocity_command_.twist =
+        controller_mapping_->ComputeVelocityCommand(joy_msg);
     velocity_command_.header.stamp = joy_msg.header.stamp;
   }
 
@@ -67,115 +83,7 @@ private:
   ros::Publisher command_pub_;
   ros::Subscriber joystick_sub_;
   geometry_msgs::TwistStamped velocity_command_;
-  const double max_linear_velocity_;
-  const double max_angular_velocity_;
-};
-
-class XboxOneControllerTeleop : public ControllerTeleop
-{
-public:
-  XboxOneControllerTeleop(
-      const ros::NodeHandle& nh, const std::string& joy_topic,
-      const std::string& command_topic, const std::string& command_frame,
-      const double max_linear_velocity, const double max_angular_velocity)
-      : ControllerTeleop(
-          nh, joy_topic, command_topic, command_frame, max_linear_velocity,
-          max_angular_velocity) {}
-
-private:
-  const size_t kLeftStickUpDown = 1;
-  const size_t kLeftStickLeftRight = 0;
-  const size_t kLeftTrigger = 2;
-  const size_t kRightTrigger = 5;
-  const size_t kAButton = 0;
-
-  geometry_msgs::Twist ComputeVelocityCommand(
-      const sensor_msgs::Joy& joy_msg) const override
-  {
-    const double left_stick_up_down = joy_msg.axes.at(kLeftStickUpDown);
-    const double left_stick_left_right = joy_msg.axes.at(kLeftStickLeftRight);
-    const double left_trigger = joy_msg.axes.at(kLeftTrigger);
-    const double right_trigger = joy_msg.axes.at(kRightTrigger);
-    const int32_t a_button = joy_msg.buttons.at(kAButton);
-    const double merged_triggers = MergeTriggers(left_trigger, right_trigger);
-
-    geometry_msgs::Twist velocity_command;
-    if (a_button == 1)
-    {
-      velocity_command.linear.x = max_linear_velocity() * left_stick_up_down;
-      velocity_command.linear.y = max_linear_velocity() * left_stick_left_right;
-      velocity_command.linear.z = 0.0;
-      velocity_command.angular.x = 0.0;
-      velocity_command.angular.y = 0.0;
-      velocity_command.angular.z = max_angular_velocity() * merged_triggers;
-    }
-    else
-    {
-      velocity_command.linear.x = 0.0;
-      velocity_command.linear.y = 0.0;
-      velocity_command.linear.z = 0.0;
-      velocity_command.angular.x = 0.0;
-      velocity_command.angular.y = 0.0;
-      velocity_command.angular.z = 0.0;
-    }
-
-    return velocity_command;
-  }
-
-  static double MergeTriggers(
-      const double left_trigger, const double right_trigger)
-  {
-    return ((-left_trigger - 1.0) * 0.5) + ((right_trigger + 1.0) * 0.5);
-  }
-};
-
-class Logitech3DProControllerTeleop : public ControllerTeleop
-{
-public:
-  Logitech3DProControllerTeleop(
-      const ros::NodeHandle& nh, const std::string& joy_topic,
-      const std::string& command_topic, const std::string& command_frame,
-      const double max_linear_velocity, const double max_angular_velocity)
-      : ControllerTeleop(
-          nh, joy_topic, command_topic, command_frame, max_linear_velocity,
-          max_angular_velocity) {}
-
-private:
-  const size_t kPitch = 1;
-  const size_t kRoll = 0;
-  const size_t kYaw = 2;
-  const size_t kThumbButton = 1;
-
-  geometry_msgs::Twist ComputeVelocityCommand(
-      const sensor_msgs::Joy& joy_msg) const override
-  {
-    const double pitch_axis = joy_msg.axes.at(kPitch);
-    const double roll_axis = joy_msg.axes.at(kRoll);
-    const double yaw_axis = joy_msg.axes.at(kYaw);
-    const int32_t thumb_button = joy_msg.buttons.at(kThumbButton);
-
-    geometry_msgs::Twist velocity_command;
-    if (thumb_button == 1)
-    {
-      velocity_command.linear.x = max_linear_velocity() * pitch_axis;
-      velocity_command.linear.y = max_linear_velocity() * roll_axis;
-      velocity_command.linear.z = 0.0;
-      velocity_command.angular.x = 0.0;
-      velocity_command.angular.y = 0.0;
-      velocity_command.angular.z = max_angular_velocity() * yaw_axis;
-    }
-    else
-    {
-      velocity_command.linear.x = 0.0;
-      velocity_command.linear.y = 0.0;
-      velocity_command.linear.z = 0.0;
-      velocity_command.angular.x = 0.0;
-      velocity_command.angular.y = 0.0;
-      velocity_command.angular.z = 0.0;
-    }
-
-    return velocity_command;
-  }
+  std::unique_ptr<ControllerMapping> controller_mapping_;
 };
 }  // namespace
 }  // namespace dji_robomaster_ep_driver
@@ -202,26 +110,10 @@ int main(int argc, char** argv)
       nhp.param(std::string("max_angular_velocity"), 10.0);
   const double publish_hz = nhp.param(std::string("publish_hz"), 30.0);
 
-  if (joystick_type == "xbox_one")
-  {
-    ROS_INFO("Creating teleop controller for Xbox One controller");
-    dji_robomaster_ep_driver::XboxOneControllerTeleop controller_teleop(
-        nh, joystick_topic, command_topic, command_frame,
-        max_linear_velocity, max_angular_velocity);
-    controller_teleop.Loop(publish_hz);
-  }
-  else if (joystick_type == "3d_pro")
-  {
-    ROS_INFO("Creating teleop controller for Logitech 3D PRO controller");
-    dji_robomaster_ep_driver::Logitech3DProControllerTeleop controller_teleop(
-        nh, joystick_topic, command_topic, command_frame,
-        max_linear_velocity, max_angular_velocity);
-    controller_teleop.Loop(publish_hz);
-  }
-  else
-  {
-    ROS_FATAL("Joystick type [%s] not supported", joystick_type.c_str());
-  }
+  dji_robomaster_ep_driver::ControllerTeleop controller_teleop(
+      nh, joystick_type, joystick_topic, command_topic,
+      command_frame, max_linear_velocity, max_angular_velocity);
+  controller_teleop.Loop(publish_hz);
 
   return 0;
 }
