+++ dji_robomaster_ep_driver/src/joystick_teleop_node.ros2.cpp
@@ -1,3 +1,6 @@
+#include <dji_robomaster_ep_driver/dji_robomaster_ep_driver.hpp>
+#include <dji_robomaster_ep_driver/joystick_controller_mappings.hpp>
+
 #include <memory>
 #include <stdexcept>
 #include <string>
@@ -9,123 +12,6 @@
 
 namespace dji_robomaster_ep_driver
 {
-class ControllerMapping
-{
-public:
-  ControllerMapping(double max_linear_velocity, double max_angular_velocity)
-    : max_linear_velocity_(max_linear_velocity),
-      max_angular_velocity_(max_angular_velocity)
-  {
-  }
-
-  virtual ~ControllerMapping() = default;
-
-  virtual geometry_msgs::msg::Twist ComputeVelocityCommand(
-      const sensor_msgs::msg::Joy& joy_msg) const = 0;
-
-  double max_linear_velocity() const { return max_linear_velocity_; }
-
-  double max_angular_velocity() const { return max_angular_velocity_; }
-
-private:
-  const double max_linear_velocity_;
-  const double max_angular_velocity_;
-};
-
-class XboxOneControllerMapping : public ControllerMapping
-{
-public:
-  using ControllerMapping::ControllerMapping;
-
-  geometry_msgs::msg::Twist ComputeVelocityCommand(
-      const sensor_msgs::msg::Joy& joy_msg) const override
-  {
-    const double left_stick_up_down = joy_msg.axes.at(kLeftStickUpDown);
-    const double left_stick_left_right = joy_msg.axes.at(kLeftStickLeftRight);
-    const double left_trigger = joy_msg.axes.at(kLeftTrigger);
-    const double right_trigger = joy_msg.axes.at(kRightTrigger);
-    const int32_t a_button = joy_msg.buttons.at(kAButton);
-    const double merged_triggers = MergeTriggers(left_trigger, right_trigger);
-
-    geometry_msgs::msg::Twist velocity_command;
-    if (a_button == 1)
-    {
-      velocity_command.linear.x = max_linear_velocity() * left_stick_up_down;
-      velocity_command.linear.y = max_linear_velocity() * left_stick_left_right;
-      velocity_command.linear.z = 0.0;
-      velocity_command.angular.x = 0.0;
-      velocity_command.angular.y = 0.0;
-      velocity_command.angular.z = max_angular_velocity() * merged_triggers;
-    }
-    else
-    {
-      velocity_command.linear.x = 0.0;
-      velocity_command.linear.y = 0.0;
-      velocity_command.linear.z = 0.0;
-      velocity_command.angular.x = 0.0;
-      velocity_command.angular.y = 0.0;
-      velocity_command.angular.z = 0.0;
-    }
-
-    return velocity_command;
-  }
-
-private:
-  const size_t kLeftStickUpDown = 1;
-  const size_t kLeftStickLeftRight = 0;
-  const size_t kLeftTrigger = 2;
-  const size_t kRightTrigger = 5;
-  const size_t kAButton = 0;
-
-  static double MergeTriggers(
-      const double left_trigger, const double right_trigger)
-  {
-    return ((-left_trigger - 1.0) * 0.5) + ((right_trigger + 1.0) * 0.5);
-  }
-};
-
-class Logitech3DProControllerMapping : public ControllerMapping
-{
-public:
-  using ControllerMapping::ControllerMapping;
-
-  geometry_msgs::msg::Twist ComputeVelocityCommand(
-      const sensor_msgs::msg::Joy& joy_msg) const override
-  {
-    const double pitch_axis = joy_msg.axes.at(kPitch);
-    const double roll_axis = joy_msg.axes.at(kRoll);
-    const double yaw_axis = joy_msg.axes.at(kYaw);
-    const int32_t thumb_button = joy_msg.buttons.at(kThumbButton);
-
-    geometry_msgs::msg::Twist velocity_command;
-    if (thumb_button == 1)
-    {
-      velocity_command.linear.x = max_linear_velocity() * pitch_axis;
-      velocity_command.linear.y = max_linear_velocity() * roll_axis;
-      velocity_command.linear.z = 0.0;
-      velocity_command.angular.x = 0.0;
-      velocity_command.angular.y = 0.0;
-      velocity_command.angular.z = max_angular_velocity() * yaw_axis;
-    }
-    else
-    {
-      velocity_command.linear.x = 0.0;
-      velocity_command.linear.y = 0.0;
-      velocity_command.linear.z = 0.0;
-      velocity_command.angular.x = 0.0;
-      velocity_command.angular.y = 0.0;
-      velocity_command.angular.z = 0.0;
-    }
-
-    return velocity_command;
-  }
-private:
-  const size_t kPitch = 1;
-  const size_t kRoll = 0;
-  const size_t kYaw = 2;
-  const size_t kThumbButton = 1;
-};
-
 class ControllerTeleopNode : public rclcpp::Node
 {
 public:
